requirements.txt
Django==1.7.1
django-crispy-forms==1.4.0
django-hstore==1.3.5
